+++ src/NetworkMessages/MapPoolRequest.cs
@@ -1,0 +1,34 @@
+﻿using System.Collections.Generic;
+using UltrakillBingoClient;
+
+namespace UltraBINGO.NetworkMessages;
+
+public class MapPoolRequest : SendMessage
+{
+    public string messageType = "GetMapPools";
+
+    public string steamId;
+    public int gameId;
+    public RegisterTicket ticket;
+}
+
+public class MapPool
+{
+    public int MapPoolId;
+    public string MapPoolName;
+    public string MapPoolDescription;
+    public int MapPoolLevelCount;
+}
+
+public class MapPoolResponse : MessageResponse
+{
+    public List<MapPool> mapPools;
+}
+
+public static class MapPoolResponseHandler
+{
+    public static void handle(MapPoolResponse response)
+    {
+        BingoMapSelection.setupMapPools(response.mapPools);
+    }
+}